Windows: AFSRedirLib AFSIgnoreReparsePointToFile
AFSIgnoreReparsePointToFile() is a helper routine that uses the
global reparse point policy to decide whether or not a reparse point
whose target is a file should be reported to applications as a file.
When per-AuthGroup or per-Process policy is supported, this function
should be modified.

Change-Id: I1e5fb5c6b901c4b56fadef4acd3c6c78f38a48d1
Reviewed-on: http://gerrit.openafs.org/9788
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
Tested-by: Jeffrey Altman <jaltman@your-file-system.com>
(cherry picked from commit f782d1dd20ee6048ca165e808adc9a01b2666746)
Reviewed-on: http://gerrit.openafs.org/9862